[cmyk-118] 채팅방, 채팅내역 API

+++ src/main/java/com/cmyk/ego/speaktoyouspring/api/hub/user_account/UserAccount.java
@@ -21,7 +21,7 @@ public class UserAccount {
 
   /// 사용자가 만든 EGO ID
   @Column(name="ego_id")
-  private Integer egoId;
+  private Long egoId;
 
   /// 사용자 email
   @Column(name="email")
